refactor: 코드 분리, 타입정의, 폴더 구조 변경 - 기존 mock데이터 및 타입 정의 정리 - 폴더 구조 및 파일명 변경 - 함수 분리 및 구조 변경 Github issue #18
