+++ package.json
@@ -7,7 +7,7 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "generate-types": "npx supabase gen types typescript --project-id knnwqvamiuvzvkrpelwz --schema public > types_db.ts"
+    "generate-types": "npx supabase gen types typescript --project-id pfhkcrkxnqrgvcuiwntt --schema public > types_db.ts"
   },
   "dependencies": {
     "@emotion/react": "^11.13.3",
