+++ src/app/page.tsx
@@ -1,12 +1,13 @@
 import { Suspense } from 'react';
 import UI from 'src/app/ui';
 import { HomeSkeleton } from './components/homeSkeleton';
-import { fetchYoutubeChannels } from '@/actions/youtubeThumbnails.action';
-import { fetchYoutubeVideos, getVideos } from '@/actions/fetchYoutubePlaylist.action';
-import { Navbar } from '@/components/navBar';
+import { fetchYoutubeChannels } from '@/actions/youtube /youtubeThumbnails.action';
+import { Navbar } from '@/components/common/navBar';
+import { getVideos } from '@/actions/youtube /video.action.ts';
+import { fetchYoutubeVideos } from '@/actions/youtube /youtube.action';
 
 export default async function PlaylistViewer() {
-  // await fetchYoutubeVideos();
+  await fetchYoutubeVideos();
   const [latestVideos, topVideo, initialChannel] = await Promise.all([
     getVideos({ sortBy: 'date' }),
     getVideos({ sortBy: 'views', limit: 5 }),
