+++ src/app/ranking/components/header/index.ts
@@ -1,0 +1,3 @@
+import Section from '@/app/ranking/components/header/header';
+
+export const Header = Object.assign(Section, {});
