--- src/app/ranking/components/headerSection/index.ts
@@ -1,3 +1,0 @@
-import Section from '@/app/ranking/components/headerSection/header';
-
-export const Header = Object.assign(Section, {});
