+++ src/app/ranking/components/videoRankingSection/card.tsx
@@ -1,5 +1,5 @@
-import { VideoLink } from '@/components/link';
-import { formatDate, formatLikeCount, formatViewCount } from '@/utils/formatters';
+import { VideoLink } from '@/components/layout /link';
+import { formatDate, formatLikeCount, formatViewCount } from '@/lib/utils/formatters';
 import Image from 'next/image';
 import { CheckCircle } from '@mui/icons-material';
