+++ src/app/ranking/components/videoRankingSection/section.tsx
@@ -17,20 +17,3 @@ export default function Section({ videos, filters, toggleMusic, selectedMusic })
     </div>
   );
 }
-
-Section.Skeleton = function Skeleton() {
-  return (
-    <div className="space-y-2">
-      {Array.from({ length: 10 }).map((_, index) => (
-        <div key={index} className="flex items-center gap-4 p-4 animate-pulse">
-          <div className="w-4 h-4 bg-gray-200 rounded" />
-          <div className="w-12 h-10 bg-gray-200 rounded" />
-          <div className="flex-1">
-            <div className="h-4 bg-gray-200 rounded w-3/4 mb-2" />
-            <div className="h-3 bg-gray-200 rounded w-1/2" />
-          </div>
-        </div>
-      ))}
-    </div>
-  );
-};
