--- src/app/ranking/loading.tsx
@@ -1,5 +1,0 @@
-import RankingSkeleton from '@/app/ranking/components/videoRankingSection/rankingSkeleton';
-
-export default function Loading() {
-  return <RankingSkeleton />;
-}
