+++ src/app/search/components/searchSection/index.tsx
@@ -1,0 +1,3 @@
+import Section from '@/app/search/components/searchSection/section';
+
+export const SearchResults = Object.assign(Section, {});
