+++ src/app/ui.tsx
@@ -1,7 +1,6 @@
 'use client';
 
 import React, { Suspense } from 'react';
-import Link from 'next/link';
 import { LatestMusicSection } from '@/app/components/latestMusicSection';
 import { TopRankingSection } from '@/app/components/topRankingSection';
 import { ChannelSection } from '@/app/components/channelSection';
