+++ src/components/common/searchInput.tsx
@@ -1,4 +1,3 @@
-// components/SearchInput.tsx
 'use client';
 
 import { ChangeEvent, useEffect, useState } from 'react';
