+++ src/hooks/useDebounce.ts
@@ -1,4 +1,3 @@
-// hooks/useDebounce.ts
 import { useEffect, useState } from 'react';
 
 export function useDebounce(value, delay: number) {
