--- src/mixins/styles.ts
@@ -1,5 +1,0 @@
-export const flexRow = 'flex flex-row';
-export const flexCol = 'flex flex-col';
-export const flexCenter = 'flex items-center justify-center';
-export const flexRowCenter = `${flexRow} ${flexCenter}`;
-export const flexColCenter = `${flexCol} ${flexCenter}`;
