--- src/mocks/youtubeData.ts
@@ -1,1245 +1,0 @@
-export const mockYoutubeData = [
-  {
-    id: 'pkypi5nXSYM',
-    viewCount: 5398878,
-    likeCount: 80590,
-    snippet: {
-      publishedAt: '2023-08-04T17:44:15Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: '아야츠노 유니 ( Ayatsuno Yuni ) | ‘내꺼 하는 법’ Music Video',
-      description:
-        'Streaming & Download : https://accel.ink/YuniYuni\n아야츠노 유니 (Ayatsuno Yuni) 1st Single 「 내꺼 하는 법 」\n\n─────────────\n\n[ Music Video Credit ]\nIllustrated by odyk\nMusic Video Directed by TEMACA\nMusic Video Assistant by  Hachie\nchoreography Designed by 박세아 / SeahPark\nArt Directed by TheCold (STELLIVE)\n\n─────────────\n\n[ Credit ]\n\nComposed by 웅킴, Stereo14\nLyrics by Ayatsuno Yuni / Lyrics Assisted by REQ\n\nVocal Directed by 웅킴, REQ\nRecorded by JiHEE at Accelers Studio\nMixed by Bongwon Shin at GLAB Studio\nEngineered for Mix by Namjun Park at GLAB Studio\nMastered by Nam-woo Kwon at 821 Sound Mastering Studio\n\nMusic Produced & Invested by Accelers, Inc.\r\nIP Supervised by STELLIVE\r\n(Music) Music Supervisor. REQ (Accelers)\r\n(IP) Executive Producer. 강지\nContents Marketing. 이로빈 (Accelers)\r\n\n© Accelers, Inc. & STELLIVE | Distributed by Dreamus Company.\r\nWARNING : All Rights Reserved. Unauthorized Duplication & Rent is Prohibited.\n\n#아야츠노유니 #내꺼하는법\n\n스텔라이브 소속 아야츠노 유니입니다☪️🤍\n---------------------------------------------------------------------------------------------------\n유니 트위치 : https://www.twitch.tv/ayatsunoyuni_stellive\n유니 트위터 : https://twitter.com/AyatsunoYuni\n유니 틱톡 : https://www.tiktok.com/@ayatsunoyuni_stellive\n유니 다시보기 채널 : https://www.youtube.com/@Yunireplay\n---------------------------------------------------------------------------------------------------\n⭐스텔라이브 \n유튜브 : https://www.youtube.com/@stellive_official\n팬카페 : https://cafe.naver.com/tteokbokk1\n트위터 : https://twitter.com/stelLive_kr',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/pkypi5nXSYM/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/pkypi5nXSYM/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/pkypi5nXSYM/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/pkypi5nXSYM/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/pkypi5nXSYM/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 23,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: 'pkypi5nXSYM',
-      },
-      videoOwnerChannelTitle: '아야츠노 유니 AYATSUNO YUNI',
-      videoOwnerChannelId: 'UClbYIn9LDbbFZ9w2shX3K0g',
-    },
-  },
-  {
-    id: 'ajDAmJYPQ-U',
-    viewCount: 4975758,
-    likeCount: 46624,
-    snippet: {
-      publishedAt: '2023-11-09T09:07:50Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: '藍璃かんな(Airi Kanna) | ‘最終花’',
-      description:
-        'Streaming & Download : https://lnk.to/TheLastFlower\n藍璃かんな (Airi Kanna) 2nd Single 「 最終花 」\n\n\nMusic Credit\n\n최종화(The Last Flower)\nLyrics by REQ, Tsubasa Harihara（HarryP）\nComposed by Tsubasa Harihara（HarryP）\nArranged by Masataka Aratomai\n\nDirected by REQ\nVocal Edited by JiHEE, 감자빵\nRecorded by JiHEE at Accelers Studio\n\nMixed by Shin Bongwon at GLAB Studios\nEngineered for Mix by Park namjun at GLAB Studios\nMastered by Kwon Namwoo at 821 Sound Mastering Studio\n─────────────\n\n[ Music Video Credit ]\nIllustrated by モ誰\nVideo Directed by Tranquillo\nAlbum Cover Designed by DIDU\nArt Directed by TheCold (STELLIVE)\n\n─────────────\n\n[ Credit ]\nMusic Produced by Accelers, Inc.\n(Music) Music Supervisor. REQ (Accelers)\n(Music) Coordination. Yumi Jeon (Accelers)\nAlbum Cover & Marketing Resources Designed by DIDU\n\n© Accelers, Inc. & STELLIVE | Distributed by Dreamus Company.\nWARNING : All Rights Reserved. Unauthorized Duplication & Rent is Prohibited.\n\n#아이리칸나 #藍璃かんな #Airi_Kanna #最終花  #the_last_flower #최종화 #stellive',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/ajDAmJYPQ-U/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/ajDAmJYPQ-U/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/ajDAmJYPQ-U/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/ajDAmJYPQ-U/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/ajDAmJYPQ-U/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 22,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: 'ajDAmJYPQ-U',
-      },
-      videoOwnerChannelTitle: '아이리 칸나 Music Official',
-      videoOwnerChannelId: 'UC6YnTqZidFg4WUiXpiCtSSQ',
-    },
-  },
-  {
-    id: 'R_RAWjqdgTs',
-    viewCount: 4075981,
-    likeCount: 70103,
-    snippet: {
-      publishedAt: '2024-06-19T09:24:43Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: '아야츠노 유니 ( Ayatsuno Yuni ) | ‘SUPADOPA’ Music Video',
-      description:
-        'Streaming & Download : https://lnk.to/SUPADOPA\n아야츠노 유니 (Ayatsuno Yuni) 2nd Single 「 SUPADOPA 」\n\n─────────────\n\n[ Music Credit ]\n\nLyrics by 아야츠노 유니, REQ\nComposed by 웅킴, Stereo14\nArranged by 웅킴, Stereo14\n\n\nVocal Directed & Edited by REQ, 웅킴, Stereo14\nRecorded by SPB at Accelers Studio\nMixed by Bongwon Shin at GLAB Studios\nEngineered for Mix by Namjun Park at GLAB Studios\nMastered by Nam-woo Kwon at 821 Sound Mastering Studio\n\n─────────────\n\nMusic Video Credit\n\nArt & Music Video Directed by TheCold (STELLIVE)\nAnimated by Jang Seokmin, blue box\nMusic Video by TEMACA\nChoreography. Suzy\n\n─────────────\n\n[ Credit ]\nMusic Produced by Accelers\n(Music) Music Supervisor. REQ (Accelers)\n(Music) Creative Assistant. 신혜경 (Accelers)\n(IP) Executive Producer. 강지 (STELLIVE)\nAlbum Cover & Marketing Resources Designed by - TheCold (STELLIVE)\n\n\n© Accelers & STELLIVE | Distributed by YG PLUS\nWARNING : All Rights Reserved. Unauthorized Duplication & Rent is Prohibited.\n\n\n#SUPADOPA #슈파도파 #유니',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/R_RAWjqdgTs/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/R_RAWjqdgTs/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/R_RAWjqdgTs/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/R_RAWjqdgTs/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/R_RAWjqdgTs/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 10,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: 'R_RAWjqdgTs',
-      },
-      videoOwnerChannelTitle: '아야츠노 유니 AYATSUNO YUNI',
-      videoOwnerChannelId: 'UClbYIn9LDbbFZ9w2shX3K0g',
-    },
-  },
-  {
-    id: 'kPdB6iGYBBc',
-    viewCount: 2945891,
-    likeCount: 32822,
-    snippet: {
-      publishedAt: '2023-05-26T09:38:53Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: '藍璃かんな(Airi Kanna) | ‘ADDICT!ON’ Music Video',
-      description:
-        'Streaming & Download : https://accel.ink/OdxJoTJE\n藍璃かんな (Airi Kanna) 1st Single 「 ADDICT!ON 」\n\n─────────────\n\n[ Music Video Credit ]\nIllustrated by pokan\nMusic Video Directed by TEMACA\nAlbum Cover Designed by TEMACA\nArt Directed by TheCold (STELLIVE)\n\n─────────────\n\n[ Credit ]\nMusic Produced & Invested by Accelers, Inc.\n(Music) Music Supervisor. REQ (Accelers)\nDigital Brand Marketing. Accelers Solutions Unit\nMarketing Resources Designed by DIDU\nPlatform Album Designed by Minirecord\n\nIP Supervised by STELLIVE\n(IP) Executive Producer. 강지 (STELLIVE)\n\n© Accelers, Inc. & STELLIVE | Distributed by Dreamus Company.\nWARNING : All Rights Reserved. Unauthorized Duplication & Rent is Prohibited.',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/kPdB6iGYBBc/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/kPdB6iGYBBc/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/kPdB6iGYBBc/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/kPdB6iGYBBc/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/kPdB6iGYBBc/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 25,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: 'kPdB6iGYBBc',
-      },
-      videoOwnerChannelTitle: '아이리 칸나 Music Official',
-      videoOwnerChannelId: 'UC6YnTqZidFg4WUiXpiCtSSQ',
-    },
-  },
-  {
-    id: 'iCCqSFvz9kM',
-    viewCount: 2030153,
-    likeCount: 20283,
-    snippet: {
-      publishedAt: '2024-03-29T11:08:01Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: '아이리 칸나(Airi Kanna) | 최종화 (Acoustic Ver.)',
-      description:
-        'Original : https://www.youtube.com/watch?v=ajDAmJYPQ-U\n\n최종화(The Last Flower) Acoustic Ver. Instrument \ndownload LINK : https://www.dropbox.com/scl/fi/2z9jp1r3a05de7i9mof06/Airi-Kanna-The-Last-Flower-Unplugged-Ver-INST.wav?rlkey=kvi2xhnt4k5gvj1mlebtsr0e5&dl=0\n\n*인스트 사용시 아래 연주자 크레딧을 필히 명시해주시길 바랍니다.\n[Acoustic Guitar. Jaehoon Jang]\n\n\n✧Director & Mix & Mastering  : REQ\n https://twitter.com/requiemdh\n\n\n✧Inst Arrange : Jaehoon Jang, REQ\n\n\n✧Acoustic Guitar. Jaehoon Jang\nhttps://www.youtube.com/@JaehoonJang\nhttps://www.instagram.com/jaehoonjang_official/\n\n\n✧Illustration/MV Director : The Cold\n  https://twitter.com/thecold_ \n\n\n✧Movie：cappu \n https://twitter.com/c4ppu \n\n\nLyrics by REQ, Tsubasa Harihara（HarryP）\nComposed by Tsubasa Harihara（HarryP）\nArranged by Jaehoon Jang, Masataka Aratomai\n\n\n✧Vocal & Chorus : Airi Kanna\nJP Twitter\n→  https://twitter.com/AiriKanna_\nKR Twitter\n→  https://twitter.com/airikannaKR\nKR Youtube \n→   https://www.youtube.com/channel/UCKzfyYWHQ92z_2jUcSABM8Q   \nKR CHZZK\n→https://chzzk.naver.com/f722959d1b8e6...',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/iCCqSFvz9kM/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/iCCqSFvz9kM/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/iCCqSFvz9kM/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/iCCqSFvz9kM/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/iCCqSFvz9kM/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 18,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: 'iCCqSFvz9kM',
-      },
-      videoOwnerChannelTitle: '아이리 칸나 Music Official',
-      videoOwnerChannelId: 'UC6YnTqZidFg4WUiXpiCtSSQ',
-    },
-  },
-  {
-    id: 'suyhkCwKX48',
-    viewCount: 1597468,
-    likeCount: 20799,
-    snippet: {
-      publishedAt: '2024-04-03T08:34:58Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: '아이리 칸나 x 아카네 리제  | ‘Hit On Shot’',
-      description:
-        'Streaming & Download : https://lnk.to/HitonShot\n아이리 칸나 (Airi Kanna) X 아카네 리제 (Akane Lize) 「 Hit On Shot 」\n\nMusic Credit\n\nLyrics by 1Take (NEWTYPE), TAK\nComposed by 1Take (NEWTYPE), TAK\nArranged by 1Take (NEWTYPE), TAK\nChorus by ARRAN\nVocal Directed by 1Take, Corbin, REQ\nKeyboard by TAK\n\nMixed by Shin Bongwon at GLAB Studios\nEngineered for Mix by Park namjun at GLAB Studios\nMastered by Kwon Namwoo at 821 Sound Mastering Studio\n\n─────────────\n\n[ Music Video Credit ]\nIllustrated by chyan\nMusic Video Directed by TEMACA\nArt Directed by TheCold (STELLIVE)\n\n─────────────\n\n[ Credit ]\nMusic Produced by Accelers\n(Music) Music Supervisor. REQ (Accelers)\n(Music) Coordination. Yumi Jeon (Accelers)\n(IP) Executive Producer. 강지 (STELLIVE)\nAlbum Cover & Marketing Resources Designed by - TEMACA\n\n© Accelers & STELLIVE | Distributed by YG PLUS\nWARNING : All Rights Reserved. Unauthorized Duplication & Rent is Prohibited.\n\n#아이리칸나 #Airi_Kanna #아카네리제 #Akane_Lize #Hit_On_Shot #stellive #스텔라이브 #힛온샷',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/suyhkCwKX48/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/suyhkCwKX48/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/suyhkCwKX48/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/suyhkCwKX48/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/suyhkCwKX48/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 16,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: 'suyhkCwKX48',
-      },
-      videoOwnerChannelTitle: '스텔라이브 StelLive Official',
-      videoOwnerChannelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-    },
-  },
-  {
-    id: 'AsoPg-h-644',
-    viewCount: 1203177,
-    likeCount: 12608,
-    snippet: {
-      publishedAt: '2023-07-10T08:58:21Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: '色彩 / Team. Palette 【ボカデュオ2023】',
-      description:
-        '“今しか見つけられない色を君と探したいんだ”\n\n\n✧色彩 / VocaDuo team. 『 Palette 』 \n\n\n✧VocaDuo team. 『 Palette 』 \n\n\n✧Vocal & Chorus : 藍璃 かんな(Airi Kanna)\n\n\n✧Lyrics : 藍璃 かんな(Airi Kanna), REQ\n✧Composed : 藍璃 かんな(Airi Kanna), REQ\n✧Arranged : REQ\n\n✧Guitar : いんいん\nhttps://youtube.com/@Oin_Guitar\n\n✧Lyrics & Rap Making : おわり-owari-\nhttps://twitter.com/owari_start\n\n✧Music Produce : REQ\nhttps://twitter.com/requiemdh\n\n✧Art Director & Produce : The Cold\nhttps://twitter.com/thecold_\n\n✧Illustration : らんぐ\nhttps://twitter.com/ran9u\n\n✧Movie：cappu\nhttps://twitter.com/c4ppu \n\n✧MIX&MASTERING ：TRMX Sound Lab \nhttps://twitter.com/Trmx_Sound_Lab\n\n\n\n\n#vocaduo2023 \n#ボカデュオ2023',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/AsoPg-h-644/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/AsoPg-h-644/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/AsoPg-h-644/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/AsoPg-h-644/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 24,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: 'AsoPg-h-644',
-      },
-      videoOwnerChannelTitle: '아이리 칸나 Music Official',
-      videoOwnerChannelId: 'UC6YnTqZidFg4WUiXpiCtSSQ',
-    },
-  },
-  {
-    id: 'gq3gzxPBOK0',
-    viewCount: 1002770,
-    likeCount: 21163,
-    snippet: {
-      publishedAt: '2024-07-24T04:34:17Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: '藍璃かんな(Airi Kanna) | ‘青い宝石と竜の子’',
-      description:
-        'Streaming & Download : https://lnk.to/Bluecrystal\n藍璃かんな(Airi Kanna) | ‘青い宝石と竜の子’\n\n[ Credit ]\n\nDirected by REQ\nVocal & Chorus by Airi Kanna\n\nLyrics by REQ\nComposed by Miyako Matsuoka\nArranged by Miyako Matsuoka \nCo-Arrangement by REQ\n\nViolin & Viola by Atsuki Yoshida (EMO Strings)\nDrums & Percussions by HAZE (Yosuke Omiya)\nPiano by Miyako Matsuoka\n\nVocal Edited by SPB at Accelers Studio\nRecorded by SPB at Accelers Studio\n\nMixed by Shin Bongwon at GLAB Studios\nEngineered for Mix by Park namjun at GLAB Studios\nMastered by Kwon Namwoo at 821 Sound Mastering Studio\n\n\nIllustrated by れおえん\nVideo Directed by TEMACA\nArt & Music Video Directed by TheCold (STELLIVE)\n\n\nMusic Produced by Accelers, Inc.\n(Music) Music Supervisor. REQ (Accelers)\n(Music) Coordination. Yumi Jeon, SPB (Accelers)\n(IP) Executive Producer. 강지 (STELLIVE)\nAlbum Cover & Marketing Resources Designed by - TEMACA\n\n© Accelers & STELLIVE | Distributed by YG PLUS\nWARNING : All Rights Reserved. Unauthorized Duplication & Rent is Prohibited.\n\n#아이리칸나 #Airi_Kanna #藍璃かんな #푸른보석과어린용  #Blue_crystal #青い宝石と竜の子 #stellive #스텔라이브',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/gq3gzxPBOK0/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/gq3gzxPBOK0/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/gq3gzxPBOK0/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/gq3gzxPBOK0/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/gq3gzxPBOK0/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 8,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: 'gq3gzxPBOK0',
-      },
-      videoOwnerChannelTitle: '아이리 칸나 Music Official',
-      videoOwnerChannelId: 'UC6YnTqZidFg4WUiXpiCtSSQ',
-    },
-  },
-  {
-    id: 'Jt6UCd7JhAA',
-    viewCount: 548779,
-    likeCount: 9497,
-    snippet: {
-      publishedAt: '2024-04-30T00:44:07Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: '스텔라이브 (StelLive) Universe | ‘Stars Align’',
-      description:
-        'Streaming & Download : https://lnk.to/Stars__Align\n스텔라이브 (StelLive) Universe 「 Stars Align 」\n\n[Music Credit]\n\nLyrics by Shirayuki Hina, Neneko Mashiro, Akane Lize, Arahashi Tabi, REQ\nComposed & Arranged by Mitsukiyo\nDirected by REQ\n\nBass by Kyunghwan Yoo\nGuitar by NezMayo\n\nDigital Edit by Jihee Park at Accelers Studio\nMixed by Jun\nMastered by Kwon Namwoo at 821 Sound Mastering Studio\n\n─────────────\n\n[ Music Video Credit ]\nIllustrated by TheCold, JW, Jang Seokmin, dustball_0, ng, Chuck, ukon60, chobi\nMusic Video Produced by Tranquillo\n\nMV & Art Directed by TheCold (STELLIVE)\n\n─────────────\n\n[ Credit ]\nMusic Produced by Accelers\n(Music) Music Supervisor. REQ (Accelers)\n(IP) Executive Producer. 강지 (STELLIVE)\nAlbum Cover & Marketing Resources Designed by - ROWOON COMPANY C&C\n, TheCold (STELLIVE)\n\n© Accelers & STELLIVE | Distributed by YG PLUS\nWARNING : All Rights Reserved. Unauthorized Duplication & Rent is Prohibited.\n\n#스텔라이브 #stellive #StarsAlign #유니버스 #Universe #네네코마시로 #아카네리제 #시라유키히나 #아라하시타비 #스타즈얼라인',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/Jt6UCd7JhAA/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/Jt6UCd7JhAA/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/Jt6UCd7JhAA/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/Jt6UCd7JhAA/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/Jt6UCd7JhAA/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 12,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: 'Jt6UCd7JhAA',
-      },
-      videoOwnerChannelTitle: '스텔라이브 StelLive Official',
-      videoOwnerChannelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-    },
-  },
-  {
-    id: 'GN-rdF8tEpU',
-    viewCount: 381130,
-    likeCount: 11064,
-    snippet: {
-      publishedAt: '2024-08-31T03:28:12Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: '스텔라이브 (StelLive) Cliché | ‘Our Tales’',
-      description:
-        "Streaming & Download : https://lnk.to/Our_Tales\n스텔라이브 (StelLive) Cliché 'Our Tales'\n\n\n「특별하지 않지만 빛나고 있는 우리들의 이야기\n\n\n\n[Music Credit] \n\nLyrics by REQ\nComposed & Arranged by sidestream \n\nGuitar by sidestream \n\nDirected by REQ (Accelers)\nDigital Edit by JUN, SPB at Accelers Studio\nMixed by Shin Bongwon at GLAB Studios\nEngineered for Mix by Park namjun at GLAB Studios\nMastered by Kwon Namwoo at 821 Sound Mastering Studio\n\n\n\n[Music Video Credit] \n\nIllustrated by Zetto, みすみ(misumi), セイクオ(Seicoh), ye_jji, リル(reel)\nMusic Video by サーナ\nMV & Art Directed by TheCold (STELLIVE)\n\n\n\n[Credit] \n\nMusic Produced by Accelers\n(Music) Music Supervisor. REQ (Accelers)\n(IP) Executive Producer. 강지 (STELLIVE)\nAlbum Cover & Marketing Resources Designed by – TheCold (STELLIVE)\n\n\n© Accelers & STELLIVE | Distributed by YG PLUS\nWARNING : All Rights Reserved. Unauthorized Duplication & Rent is Prohibited. \n\n#스텔라이브 #stellive #OurTales #클리셰 #Cliché #텐코시부키 #아오쿠모린 #하나코나나 #유즈하리코",
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/GN-rdF8tEpU/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/GN-rdF8tEpU/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/GN-rdF8tEpU/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/GN-rdF8tEpU/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/GN-rdF8tEpU/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 6,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: 'GN-rdF8tEpU',
-      },
-      videoOwnerChannelTitle: '스텔라이브 StelLive Official',
-      videoOwnerChannelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-    },
-  },
-  {
-    id: 'rQaluJS-Tc0',
-    viewCount: 340803,
-    likeCount: 11297,
-    snippet: {
-      publishedAt: '2024-09-11T03:20:43Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: '시라유키 히나(Shirayuki Hina) | ‘하나(夏拏)’',
-      description:
-        'Streaming & Download : https://lnk.to/CatchtheSummer\n\n시라유키 히나 (Shirayuki Hina) 1st Single 「 하나(夏拏) 」\n\nJP Lyrics \n誰もいない線路の上を　羽ばたいてく翼は\nどこを目指しているの　教えてほしかった\n\n誰かの足跡をなぞって　歩くのが好きだった\n夏に溶かされて\n\n置いてかないで　一人にしないで　それが答えなんでしょ\nこのままきっと　誤魔化しそっと　生きていこうだなんてさ\nずるいよそんな　負け方ばっか　考えてたらもう\nいつか夢見て歌った唄も　忘れちゃうんだよ\n口ずさんでいたあの日を\n\n何も知らない未来に　怯えてしまうならいっそ\nほら　飛び込んでみようかって\n海の匂いに僕らいつでも　誘われて飛び込んだ先で\n暗い世界を　かき分けて\n光る星を見つけたんだ　そこにしかなかったんだ\n\n誰かの声を聴きながら　眠るのが好きだった\n夏に溶かされて\n\n君も僕も　同じなんだ　特別なんかなくて\n怖いよきっと　それでもずっと\n笑っていてくれるから　ださいよこんな　逃げ方ばっか\n考えてたらもう　いつか傍にいる人を傷つけてしまうんだよ\n信じてみたいあの日を\n\n何も見えない未来を　疑がってしまうならいっそ\nほら　さらけ出してみようかって\n空の色に僕らいつでも　憧れ手を伸ばしたんだ\n広い世界を　迷いながら\n光る星をまた探していた\n\nこのまま変わらず\n大人になっていくと思ってた\n日陰はもう　形を変えてゆく\nどこへ向かおうか\n\n何も知らない未来に　怯えてしまうならいっそ\nほら　飛び込んでみようかって\n海の匂いに　夏の匂いに\n僕らいつでも　誘われて飛び込んだ先で\n暗い世界を　かき分けて\n光る星を見つけたように　僕だけの星があるのなら\nこれから迎えに行くんだ\n\n\n─────────────\n[Music Credit]\n\nComposed & Arranged by *Luna\nJapanese Lyrics *Luna\nKorean Lyrics by REQ\n\nDirected by REQ (Accelers)\nDigital Edit by JUN, SPB at Accelers Studio\nMixed by *Luna\nMastered by *Luna\n─────────────\n\n[ Music Video Credit ]\n\nIllustrated by pong, TheCold\nMusic Video by Tranquillio\n\nMV & Art Directed by TheCold (STELLIVE)\n\n─────────────\n\n[ Credit ]\nMusic Produced by Accelers\n(Music) Music Supervisor. REQ (Accelers)\n(IP) Executive Producer. 강지 (STELLIVE)\nAlbum Cover & Marketing Resources Designed by – TheCold (STELLIVE)\n\n© Accelers & STELLIVE | Distributed by YG PLUS\nWARNING : All Rights Reserved. Unauthorized Duplication & Rent is Prohibited.\n\n\n#스텔라이브 #시라유키히나 #히나 #Shirayuki_Hina #하나 #夏拏',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/rQaluJS-Tc0/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/rQaluJS-Tc0/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/rQaluJS-Tc0/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/rQaluJS-Tc0/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/rQaluJS-Tc0/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 4,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: 'rQaluJS-Tc0',
-      },
-      videoOwnerChannelTitle: '시라유키 히나 SHIRAYUKI HINA',
-      videoOwnerChannelId: 'UC1afpiIuBDcjYlmruAa0HiA',
-    },
-  },
-  {
-    id: '_8uoathMY0A',
-    viewCount: 301747,
-    likeCount: 12492,
-    snippet: {
-      publishedAt: '2024-10-16T09:04:30Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: '네네코 마시로(Neneko Mashiro) | ‘봄꿈’',
-      description:
-        'Streaming & Download : https://lnk.to/Springdream\n네네코 마시로(Neneko Mashiro) | ‘봄꿈’\n\n─────────────\n[Music Credit]\n\nComposed by 김민철, 하야테\nArranged by 김민철, 하야테, 유상우, REQ\nLyrics by 김민철, 하야테\nChorus by RAY\n\nDirected by REQ (Accelers)\nVocal Edited by JUN, SPB at Accelers Studio\n\nMixed by Shin Bongwon at GLAB Studios\nEngineered for Mix by Park namjun at GLAB Studios\nMastered by Kwon Namwoo at 821 Sound Mastering Studio\n─────────────\n\n[ Music Video Credit ]\n\nArt & Music Video Directed by TheCold (STELLIVE) \nAnimated & Character Design by 파랑상자 (Blue Box)\nBackground by 파랑상자 (Blue Box)\nMusic Video by HP\n\n─────────────\n\n[ Credit ]\nMusic Produced by Accelers\n(Music) Music Supervisor. REQ (Accelers)\n(IP) Executive Producer. 강지 (STELLIVE)\nAlbum Cover & Marketing Resources Designed by – TheCold (STELLIVE)\n\n© Accelers & STELLIVE | Distributed by YG PLUS\nWARNING : All Rights Reserved. Unauthorized Duplication & Rent is Prohibited.\n\n─────────────\n\n🥛Neneko Mashiro🥛\n\n🧇Twitter l https://x.com/NenekoMashiro\n🧇CHZZK l https://chzzk.me/EXCZx\n🧇YOUTUBE l https://www.youtube.com/channel/UC_eeSpMBz8PG4ssdBPnP07g\n🧇다시보기 l https://www.youtube.com/channel/UCW3wobzj-JJE_TNs7ysJPiQ\n\n#네네코마시로 #스텔라이브 #봄꿈 #Springdream #Neneko_Mashiro #StelLive',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/_8uoathMY0A/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/_8uoathMY0A/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/_8uoathMY0A/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/_8uoathMY0A/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/_8uoathMY0A/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 0,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: '_8uoathMY0A',
-      },
-      videoOwnerChannelTitle: '네네코 마시로 NENEKO MASHIRO',
-      videoOwnerChannelId: 'UC_eeSpMBz8PG4ssdBPnP07g',
-    },
-  },
-  {
-    id: '3_0Ut2tlxLM',
-    viewCount: 301505,
-    likeCount: 11344,
-    snippet: {
-      publishedAt: '2024-09-29T06:21:17Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: '아카네 리제(Akane Lize) | ‘Festa!’',
-      description:
-        'Streaming & Download : https://lnk.to/Lize_Festa\n아카네 리제(Akane Lize) – Festa!\n\n─────────────\n[Music Credit]\n\nComposed by MooF (MonoTree), Ashley Alisha\nArranged by MooF (MonoTree), REQ\nLyrics by Akane Lize, REQ\n\nPiano by MooF\nBass by MooF\nSynthesizer by MooF\nGuitar by I.vin, Woong\n\n\nDirected by REQ (Accelers)\nVocal Edited by JUN, SPB at Accelers Studio\n\nMixed by Shin Bongwon at GLAB Studios\nEngineered for Mix by Park namjun at GLAB Studios\nMastered by Pread Choi at Accelers Studio \n─────────────\n\n[ Music Video Credit ]\n\nIllustrated by らんぐ(ran9u), DANDAN\nMusic Video by cappu\n\nMV & Art Directed by TheCold (STELLIVE)\n\n─────────────\n\n[ Credit ]\nMusic Produced by Accelers\n(Music) Music Supervisor. REQ (Accelers)\n(IP) Executive Producer. 강지 (STELLIVE)\nAlbum Cover & Marketing Resources Designed by – TheCold (STELLIVE)\n\n© Accelers & STELLIVE | Distributed by YG PLUS\nWARNING : All Rights Reserved. Unauthorized Duplication & Rent is Prohibited.\n\n#아카네리제 #akanelize #Festa! #stellive #스텔라이브',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/3_0Ut2tlxLM/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/3_0Ut2tlxLM/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/3_0Ut2tlxLM/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/3_0Ut2tlxLM/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/3_0Ut2tlxLM/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 2,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: '3_0Ut2tlxLM',
-      },
-      videoOwnerChannelTitle: '아카네 리제 AKANE LIZE',
-      videoOwnerChannelId: 'UC7-m6jQLinZQWIbwm9W-1iw',
-    },
-  },
-  {
-    id: 'YcyDo-jp4YM',
-    viewCount: 270254,
-    likeCount: 7193,
-    snippet: {
-      publishedAt: '2024-02-23T12:03:42Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: 'STELLIVE (스텔라이브) ‘Milky Way’ Music Video',
-      description:
-        "STELLIVE (스텔라이브) 1st Single 'Milky Way' \n🎵 노래듣기 : https://lnk.to/STELLIVE_MilkyWay\n\nMilky Way\n\nLyrics by 박지연(MonoTree)\nComposed by pdly, 권애진(MonoTree)\nArranged by pdly(MonoTree)\n\nBackground Vocals by 권애진\nKeyboard performed by pdly\nSynthesizer performed by pdly\nDrum programmed by pdly\nVocal directed by 권애진, pdly\n\nㅡ\n\nCredit\n\nMusic Produced by Accelers\nMusic Supervisor. REQ (Accelers)\nAlbum Cover & Marketing Resources Designed by DIDU\n\nProduction SJGS Studio,Inc.\n3D Engineering EMP Inc.\n\n\nIP Supervised by STELLIVE\nExecutive Producer. 강지 (STELLIVE)\nArt Directed by TheCold (STELLIVE)\n\n© Accelers & STELLIVE | Distributed by Dreamus Company.\nWARNING : All Rights Reserved. Unauthorized Duplication & Rent is Prohibited.\n------------------------------------------------------------------------------------------------\n\n🌟팬카페 : https://cafe.naver.com/tteokbokk1\n✉️ 공식 트위터: https://twitter.com/StelLive_kr\n✉️ 공식 홈페이지: https://stellive.me/\n\n------------------------------------------------------------------------------------------------\n\n✦아이리 칸나 : https://chzzk.naver.com/f722959d1b8e651bd56209b343932c01\n✦아야츠노 유니 : https://chzzk.naver.com/45e71a76e949e16a34764deb962f9d9f\n\n------------------------------------------------------------------------------------------------\n\n✦네네코 마시로 : https://chzzk.naver.com/4515b179f86b67b4981e16190817c580\n✦시라유키 히나 : https://chzzk.naver.com/b044e3a3b9259246bc92e863e7d3f3b8\n✦아라하시 타비 : https://chzzk.naver.com/a6c4ddb09cdb160478996007bff35296\n✦아카네 리제 : https://chzzk.naver.com/4325b1d5bbc321fad3042306646e2e50",
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/YcyDo-jp4YM/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/YcyDo-jp4YM/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/YcyDo-jp4YM/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/YcyDo-jp4YM/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/YcyDo-jp4YM/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 19,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: 'YcyDo-jp4YM',
-      },
-      videoOwnerChannelTitle: '스텔라이브 StelLive Official',
-      videoOwnerChannelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-    },
-  },
-  {
-    id: 'uTj88854gZ0',
-    viewCount: 234494,
-    likeCount: 6002,
-    snippet: {
-      publishedAt: '2024-04-20T15:03:45Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: '시라유키 히나 X 네네코 마시로 | ‘Dear my fairy’',
-      description:
-        'Streaming & Download : https://lnk.to/DearMyFairy\n시라유키 히나 (Shirayuki Hina) X 네네코 마시로 (Neneko Mashiro) 「 Dear My Fairy 」\n\n─────────────\n\nMusic Credit\n\nLyrics by whyminsu\nComposed by TAK, whyminsu\nArranged by TAK, whyminsu\nChorus by Lily\nVocal Directed by REQ, TAK\nKeyboard by TAK\nGuitar by 문원우\n\nMixed by TAK\nMastered by Kwon Namwoo at 821 Sound Mastering Studio\n\n─────────────\n\n[ Music Video Credit ]\nIllustrated by U\nMusic Video Directed by awa\nArt Directed by TheCold (STELLIVE)\n\n─────────────\n\n[ Credit ]\nMusic Produced by Accelers\n(Music) Music Supervisor. REQ (Accelers)\n(IP) Executive Producer. 강지 (STELLIVE)\nAlbum Cover & Marketing Resources Designed by - TEMACA\n\n© Accelers & STELLIVE | Distributed by YG PLUS\nWARNING : All Rights Reserved. Unauthorized Duplication & Rent is Prohibited.\n\n#네네코마시로 #Neneko_Mashiro #시라유키히나 #Shirayuki_Hina #Dear_my_fairy #stellive #스텔라이브 #디어마이페어리',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/uTj88854gZ0/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/uTj88854gZ0/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/uTj88854gZ0/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/uTj88854gZ0/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/uTj88854gZ0/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 14,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: 'uTj88854gZ0',
-      },
-      videoOwnerChannelTitle: '스텔라이브 StelLive Official',
-      videoOwnerChannelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-    },
-  },
-  {
-    id: 'Fy-GoSLrTOo',
-    viewCount: 167720,
-    likeCount: 3885,
-    snippet: {
-      publishedAt: '2023-12-02T07:12:24Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: '스텔라이브 Music Project 001',
-      description:
-        '스텔라이브의 첫 뮤직 프로젝트 티저 공개.\nComing in 2024\n\n© Accelers & STELLIVE\n\n\n\n\n\n📺 공식 트위치: https://www.twitch.tv/stellive_official\n🌟팬카페 : https://cafe.naver.com/tteokbokk1\n✉️ 공식 트위터: https://twitter.com/StelLive_kr\n\n------------------------------------------------------------------------------------------------\n✦아이리 칸나 : https://www.twitch.tv/airikanna_stellive\n✦아야츠노 유니 : https://www.twitch.tv/ayatsunoyuni_stellive\n\n✦네네코 마시로 : https://www.twitch.tv/nenekomashiro_stellive\n✦시라유키 히나 : https://www.twitch.tv/shirayukihina_stellive\n✦아라하시 타비 : https://www.twitch.tv/arahashitabi_stellive\n✦아카네 리제 : https://www.twitch.tv/akanelize_stellive',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/Fy-GoSLrTOo/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/Fy-GoSLrTOo/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/Fy-GoSLrTOo/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/Fy-GoSLrTOo/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/Fy-GoSLrTOo/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 21,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: 'Fy-GoSLrTOo',
-      },
-      videoOwnerChannelTitle: '스텔라이브 StelLive Official',
-      videoOwnerChannelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-    },
-  },
-  {
-    id: '3eZGQ_Fqqko',
-    viewCount: 153720,
-    likeCount: 4118,
-    snippet: {
-      publishedAt: '2024-02-20T06:12:51Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: 'STELLIVE (스텔라이브) ‘Milky Way’ TEASER',
-      description:
-        "STELLIVE (스텔라이브) 1st Single 'Milky Way' \n\n🎵노래듣기 : https://lnk.to/STELLIVE_MilkyWay\n\n[음원] 2024년 2월 21일 오후 6시 \n[MV] 2024년 2월 24일 오후 6시 30분\n\n스텔라이브의 첫 그룹 싱글 Milky Way가\n2024년 2월 21일 (수요일) 오후 6시 발매됩니다.",
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/3eZGQ_Fqqko/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/3eZGQ_Fqqko/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/3eZGQ_Fqqko/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/3eZGQ_Fqqko/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/3eZGQ_Fqqko/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 20,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: '3eZGQ_Fqqko',
-      },
-      videoOwnerChannelTitle: '스텔라이브 StelLive Official',
-      videoOwnerChannelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-    },
-  },
-  {
-    id: 'pU6EL5Purlc',
-    viewCount: 123498,
-    likeCount: 3968,
-    snippet: {
-      publishedAt: '2024-06-17T04:31:22Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: '아야츠노 유니 | ‘SUPADOPA’ Teaser',
-      description:
-        '아야츠노 유니 (Ayatsuno Yuni) 「 SUPADOPA 」\n2024년 6월 19일 (수요일) 오후 6시 발매!\n\n_\n\n[Credit]\nArt & Music Video Directed by TheCold (STELLIVE)\nAnimated by Jang Seokmin, blue box\nMusic Video by TEMACA\nChoreography. Suzy\n© Accelers, Inc. & STELLIVE\n\n#SUPADOPA #아야츠노_유니 #유니 #Ayatsuno_Yuni #stellive #스텔라이브',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/pU6EL5Purlc/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/pU6EL5Purlc/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/pU6EL5Purlc/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/pU6EL5Purlc/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/pU6EL5Purlc/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 11,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: 'pU6EL5Purlc',
-      },
-      videoOwnerChannelTitle: '아야츠노 유니 AYATSUNO YUNI',
-      videoOwnerChannelId: 'UClbYIn9LDbbFZ9w2shX3K0g',
-    },
-  },
-  {
-    id: 'TcWU55coMK8',
-    viewCount: 80142,
-    likeCount: 3681,
-    snippet: {
-      publishedAt: '2024-04-01T15:23:07Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: 'Airi Kanna x Akane Lize  | ‘Hit On Shot’ Teaser',
-      description:
-        '아이리 칸나 (Airi Kanna) X 아카네 리제 (Akane Lize) 「 Hit On Shot 」\n2024년 4월 3일 (수요일) 오후 6시 발매!\n\n_\n\n[Credit]\nIllustrated by Chyan\nVideo Directed by TEMACA\nArt Directed by TheCold (STELLIVE)\n© Accelers, Inc. & STELLIVE\n\n#아이리칸나 #Airi_Kanna #아카네리제 #Akane_Lize #Hit_On_Shot #stellive #스텔라이브',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/TcWU55coMK8/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/TcWU55coMK8/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/TcWU55coMK8/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/TcWU55coMK8/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/TcWU55coMK8/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 17,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: 'TcWU55coMK8',
-      },
-      videoOwnerChannelTitle: '스텔라이브 StelLive Official',
-      videoOwnerChannelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-    },
-  },
-  {
-    id: 'tZHAWd6yIz4',
-    viewCount: 78602,
-    likeCount: 4993,
-    snippet: {
-      publishedAt: '2024-07-24T04:34:33Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: '아이리 칸나(Airi Kanna) | ‘푸른 보석과 어린 용’ Teaser',
-      description:
-        '2024년 7월 25일 오후 6시, 국내 및 글로벌 음원사이트 발매 예정\n\n아이리 칸나 (Airi Kanna) 3rd Single 「 푸른 보석과 어린 용 」\n2024년 7월 25일 (목요일) 오후 6시 발매!\n\n_\n\n[Credit]\nIllustrated by れおえん ( reoenl )\nVideo Directed by TEMACA\nArt Directed by TheCold (STELLIVE)\n© Accelers, Inc. & STELLIVE\n\n#아이리칸나 #Airi_Kanna #藍璃かんな #푸른보석과어린용  #Blue_crystal #青い宝石と竜の子 #stellive #스텔라이브',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/tZHAWd6yIz4/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/tZHAWd6yIz4/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/tZHAWd6yIz4/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/tZHAWd6yIz4/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/tZHAWd6yIz4/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 9,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: 'tZHAWd6yIz4',
-      },
-      videoOwnerChannelTitle: '아이리 칸나 Music Official',
-      videoOwnerChannelId: 'UC6YnTqZidFg4WUiXpiCtSSQ',
-    },
-  },
-  {
-    id: 'xXUA_kSAp0A',
-    viewCount: 40483,
-    likeCount: 2666,
-    snippet: {
-      publishedAt: '2024-10-15T09:01:30Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: '네네코 마시로(Neneko Mashiro) | ‘봄꿈’ Teaser',
-      description:
-        '#네네코마시로 #스텔라이브 #봄꿈 #Springdream #Neneko_Mashiro #StelLive\n\n네네코 마시로(Neneko Mashiro) 1st Single 「 봄꿈 」\n2024년 10월 16일 (수요일) 오후 6시 발매\n\n🥛Neneko Mashiro🥛\n\n🧇Twitter l https://x.com/NenekoMashiro\n🧇CHZZK l https://chzzk.me/EXCZx\n🧇YOUTUBE l https://www.youtube.com/channel/UC_eeSpMBz8PG4ssdBPnP07g\n🧇다시보기 l https://www.youtube.com/channel/UCW3wobzj-JJE_TNs7ysJPiQ\n_\n\n[Credit]\nAnimated & Background by  파랑상자(Blue Box)\nMusic Video by HP\nMV & Art Directed by TheCold (STELLIVE)\n\nⓒ 2024 STELLIVE. All Rights Reserved',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/xXUA_kSAp0A/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/xXUA_kSAp0A/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/xXUA_kSAp0A/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/xXUA_kSAp0A/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/xXUA_kSAp0A/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 1,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: 'xXUA_kSAp0A',
-      },
-      videoOwnerChannelTitle: '네네코 마시로 NENEKO MASHIRO',
-      videoOwnerChannelId: 'UC_eeSpMBz8PG4ssdBPnP07g',
-    },
-  },
-  {
-    id: 'zRNkCpCQ8Sg',
-    viewCount: 38310,
-    likeCount: 2554,
-    snippet: {
-      publishedAt: '2024-09-28T06:07:03Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: '아카네 리제(Akane Lize) | ‘Festa!’ Teaser',
-      description:
-        '2024년 9월 29일 오후 6시, 국내 및 글로벌 음원사이트 발매 예정\n\n아카네 리제(Akane Lize) 1st Single 「 Festa! 」\n2024년 9월 29일 (일요일) 오후 6시 발매!\n\n_\n\n[Credit]\nIllustrated by  DANDAN, ran9u\nVideo Directed by cappu\nArt Directed by TheCold (STELLIVE)\n© Accelers, Inc. & STELLIVE',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/zRNkCpCQ8Sg/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/zRNkCpCQ8Sg/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/zRNkCpCQ8Sg/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/zRNkCpCQ8Sg/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/zRNkCpCQ8Sg/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 3,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: 'zRNkCpCQ8Sg',
-      },
-      videoOwnerChannelTitle: '아카네 리제 AKANE LIZE',
-      videoOwnerChannelId: 'UC7-m6jQLinZQWIbwm9W-1iw',
-    },
-  },
-  {
-    id: 'd7jeeCnLvRE',
-    viewCount: 34746,
-    likeCount: 2318,
-    snippet: {
-      publishedAt: '2024-08-23T03:00:40Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: '스텔라이브 (StelLive) Cliché | ‘Our Tales’ Teaser',
-      description:
-        'StelLive Cliché 「 Our Tales 」\n\n\n[Credit]\nIllustrated by Zetto みすみ(misumi), セイクオ(Seicoh), ye_jji, リル(reel)\nMusic Video Produced by サーナ\nMV & Art Directed by TheCold (STELLIVE)\n© Accelers, Inc. & STELLIVE\n\n#스텔라이브 #stellive #OurTales #클리셰 #Cliché #텐코시부키 #아오쿠모린 #하나코나나 #유즈하리코',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/d7jeeCnLvRE/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/d7jeeCnLvRE/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/d7jeeCnLvRE/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/d7jeeCnLvRE/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/d7jeeCnLvRE/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 7,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: 'd7jeeCnLvRE',
-      },
-      videoOwnerChannelTitle: '스텔라이브 StelLive Official',
-      videoOwnerChannelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-    },
-  },
-  {
-    id: '7C9r8E7wxPA',
-    viewCount: 34205,
-    likeCount: 2047,
-    snippet: {
-      publishedAt: '2024-04-29T01:39:44Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: 'StelLive Universe | ‘Stars Align’ Teaser',
-      description:
-        'StelLive Universe 「 Stars Align 」\n2024년 4월 30일 (화요일) 오후 12시 발매!\n\n[Credit]\nIllustrated by TheCold, JW, Jang Seokmin, dustball_0, ng, Chuck, ukon60, chobi \nMusic Video Produced by Tranquillo\nMV & Art Directed by TheCold (STELLIVE)\n© Accelers, Inc. & STELLIVE\n\n#스텔라이브 #stellive #StarsAlign #유니버스 #Universe #네네코마시로 #아카네리제 #시라유키히나 #아라하시타비',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/7C9r8E7wxPA/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/7C9r8E7wxPA/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/7C9r8E7wxPA/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/7C9r8E7wxPA/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/7C9r8E7wxPA/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 13,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: '7C9r8E7wxPA',
-      },
-      videoOwnerChannelTitle: '스텔라이브 StelLive Official',
-      videoOwnerChannelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-    },
-  },
-  {
-    id: 'h47SrZ9x6UA',
-    viewCount: 30758,
-    likeCount: 2444,
-    snippet: {
-      publishedAt: '2024-09-11T03:20:36Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: '시라유키 히나(Shirayuki Hina) | ‘하나(夏拏)’ Teaser',
-      description:
-        '2024년 9월 11일 오후 12시, 국내 및 글로벌 음원사이트 발매 예정\n\n시라유키 히나 (Shirayuki Hina) 1st Single 「 하나(夏拏) 」\n2024년 9월 11일 (수요일) 오후 12시 발매!\n\n_\n\n[Credit]\nIllustrated by pong\nVideo Directed by Tranquillo\nArt Directed by TheCold (STELLIVE)\n© Accelers, Inc. & STELLIVE\n\n#스텔라이브 #시라유키 #히나 #Shirayuki_Hina #하나 #夏拏',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/h47SrZ9x6UA/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/h47SrZ9x6UA/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/h47SrZ9x6UA/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/h47SrZ9x6UA/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/h47SrZ9x6UA/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 5,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: 'h47SrZ9x6UA',
-      },
-      videoOwnerChannelTitle: '시라유키 히나 SHIRAYUKI HINA',
-      videoOwnerChannelId: 'UC1afpiIuBDcjYlmruAa0HiA',
-    },
-  },
-  {
-    id: 'thTCJcevk20',
-    viewCount: 20564,
-    likeCount: 1329,
-    snippet: {
-      publishedAt: '2024-04-19T16:09:25Z',
-      channelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-      title: 'Shirayuki Hina X Neneko Mashiro | ‘Dear my fairy’ Teaser',
-      description:
-        '시라유키 히나 (Shirayuki Hina) X 네네코 마시로 (Neneko Mashiro) 「 Dear my fairy 」\n2024년 4월 21일 (일요일) 오후 12시 발매!\n\n_\n\n[Credit]\nIllustrated by U\nVideo Directed by awa\nArt Directed by TheCold (STELLIVE)\n© Accelers, Inc. & STELLIVE\n\n#네네코마시로 #Neneko_Mashiro #시라유키히나 #Shirayuki_Hina #Dear_my_fairy #stellive #스텔라이브 #디어마이페어리',
-      thumbnails: {
-        default: {
-          url: 'https://i.ytimg.com/vi/thTCJcevk20/default.jpg',
-          width: 120,
-          height: 90,
-        },
-        medium: {
-          url: 'https://i.ytimg.com/vi/thTCJcevk20/mqdefault.jpg',
-          width: 320,
-          height: 180,
-        },
-        high: {
-          url: 'https://i.ytimg.com/vi/thTCJcevk20/hqdefault.jpg',
-          width: 480,
-          height: 360,
-        },
-        standard: {
-          url: 'https://i.ytimg.com/vi/thTCJcevk20/sddefault.jpg',
-          width: 640,
-          height: 480,
-        },
-        maxres: {
-          url: 'https://i.ytimg.com/vi/thTCJcevk20/maxresdefault.jpg',
-          width: 1280,
-          height: 720,
-        },
-      },
-      channelTitle: '스텔라이브 StelLive Official',
-      playlistId: 'PLLjd981H8qSMGC4Nir0hD2Gj9n9PDUoHX',
-      position: 15,
-      resourceId: {
-        kind: 'youtube#video',
-        videoId: 'thTCJcevk20',
-      },
-      videoOwnerChannelTitle: '스텔라이브 StelLive Official',
-      videoOwnerChannelId: 'UC2b4WRE5BZ6SIUWBeJU8rwg',
-    },
-  },
-];
